backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apscheduler
attrs==25.3.0
bcrypt==4.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
cssselect2==0.8.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.4.3
openai==0.28.1
passlib[bcrypt]==1.7.4
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.0
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
sentence-transformers
setuptools==65.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.22.0
weasyprint==65.1
webencodings==0.5.1
yarl==1.20.0
zopfli==0.2.3.post1

# Ethereum / SIWE dependencies
eth-account>=0.13.1,<0.14
ckzg>=2.0.0
web3>=7.3.0,<8.0.0

# SIWE (Sign-In-With-Ethereum)
siwe @ git+https://github.com/spruceid/siwe-py.git@v4.4.0#egg=siwe
alembic==1.16.1
psycopg2-binary==2.9.10
